Only redefine ColorColumn if g:docile_color_column is set
Currently vim-docile will always redefine the highlight group for the
colocolumn to red unless specified otherwise. This conflicts with other
settings that set the highlight group of `ColorColumn` to another value.

As red is the default color for the colorcolumn anyway and any user
wanting to set it to a different color for vim-docile by defining
`g:docile_color_column` the `ColorColumn` highlight group is not only
modified if explicitly requested by the user.

Note however, that issue
https://github.com/tandrewnichols/vim-docile/issues/1 still applies if
`g:docile_color_column` is set.